feat(web): show avatar if logged in
